Filter out unclaimable pitches
