+++ client/src/pages/pitchDoc/index.tsx
@@ -1,5 +1,5 @@
 import { isEqual, startsWith, toLower, toString } from 'lodash';
-import React, { ReactElement, useEffect, useState } from 'react';
+import React, { ReactElement, useCallback, useEffect, useState } from 'react';
 import { Input, Menu } from 'semantic-ui-react';
 import { IPitch } from 'ssw-common';
 
@@ -22,6 +22,7 @@ import {
 } from '../../components';
 import { useAuth } from '../../contexts';
 import { pagesEnum } from '../../utils/enums';
+import { getClaimableTeams } from '../../utils/helpers';
 
 import { filterInterests, filterClaimStatus, filterTeams } from './helpers';
 
@@ -50,7 +51,7 @@ const PitchDoc = (): ReactElement => {
   const [teams, setTeams] = useState<string[]>([]);
   const [query, setQuery] = useState('');
 
-  const { isAdmin, isStaff } = useAuth();
+  const { user, isAdmin, isStaff } = useAuth();
 
   const getApproved = async (): Promise<void> => {
     const res = await getApprovedPitches();
@@ -60,15 +61,21 @@ const PitchDoc = (): ReactElement => {
     }
   };
 
-  const getUnclaimed = async (): Promise<void> => {
+  const getUnclaimed = useCallback(async (): Promise<void> => {
     const res = await getUnclaimedPitches();
 
     if (!isError(res)) {
-      setUnclaimed(res.data.result);
-      setCurrentPitches(res.data.result);
-      setFilteredPitches(res.data.result);
+      const unclaimedPitches = res.data.result;
+      const claimablePitches = unclaimedPitches.filter(
+        (pitch: IPitch) => getClaimableTeams(pitch, user).length > 0,
+      );
+
+
+      setUnclaimed(claimablePitches);
+      setCurrentPitches(claimablePitches);
+      setFilteredPitches(claimablePitches);
     }
-  };
+  }, [user]);
 
   const getPendingApprovals = async (): Promise<void> => {
     const res = await getPitchesPendingApproval();
@@ -107,7 +114,7 @@ const PitchDoc = (): ReactElement => {
       setFilteredPitches([]);
       setCurrentPitches([]);
     };
-  }, [isAdmin, isStaff]);
+  }, [isAdmin, isStaff, getUnclaimed]);
 
   useEffect(() => {
     const search = (pitches: IPitch[]): IPitch[] => {
