+++ client/src/utils/helpers.ts
@@ -49,6 +49,20 @@ const getUserFullName = (user: Partial<IUser>): string =>
   }`;
 
 /**
+ * Gets the teams on a pitch that a user can claimed
+ *
+ * @param pitch the pitch to check the teams for
+ * @param user the user to check the teams for
+ * @returns the teams that the user can claim
+ */
+const getClaimableTeams = (pitch: IPitch, user: IUser): string[] =>
+  pitch.writer
+    ? pitch.teams
+        .filter((team) => team.target > 0 && user.teams.includes(team.teamId))
+        .map((team) => team.teamId)
+    : [];
+
+/**
  * Parses an array of options into Semantic UI style Dropdown Items objects
  *
  * @param options the dropdown options to create
@@ -137,6 +151,7 @@ export {
   getPitchTeams,
   updateUserField,
   getUserFullName,
+  getClaimableTeams,
   parseOptions,
   isPitchClaimed,
   convertMap,
